Make application data reactive to date
The main change is that we can't simply reassign the observables,
because consumers would still be connected to the old ones. So, we need
to change those to behavior subjects: the object stays the same but new
values come in.

Another tricky part is the rounds config is expected to be an object
with useful field values (vs returning an empty array of round configs).
It would be ideal to clean that up to work like weeks, we just fetch out
the field we actually need. For now, just return a dummy object.