Let the administrator set their own password
On the admin password panel, add an "administrator" line for the admin
to change their own password.

Changing other users' passwords goes through our backend (to use the
admin SDK). But admin passwords can be changed by the client as usual,
because we're changing the logged-in  user.

This does not support reauthenticating the user if they haven't logged
in recently enough. We'll need to see what the error experience is like.

Fixes #85